Resolving merge conflicts and fixing switch statements
